% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_data.R
\name{analysis_data}
\alias{analysis_data}
\title{Study datatable}
\usage{
analysis_data(analysis_id, untarg = FALSE)
}
\arguments{
\item{analysis_id}{the id of a study e.g. "AN000001".}
}
\value{
a data frame
}
\description{
Get the measured metabolite values for an analysis (NB may have been
normalised etc see analysis for details). Does not return class labels as per
api documentation; use `study_factors`
}
