% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_def.R
\name{check_pattern}
\alias{check_pattern}
\alias{check_pattern,mw_input_item-method}
\title{Check input against acceptable input pattern}
\usage{
check_pattern(I, input_value, match)

\S4method{check_pattern}{mw_input_item}(I, input_value, match)
}
\arguments{
\item{I}{An mw_input_item.}

\item{input_value}{The value for the input item (character).}

\item{match}{The type of match. One of "exact" or "partial".}
}
\value{
TRUE if input matches the pattern, or throws an error.
}
\description{
Checks an input_value against a regex pattern to determine if the input_value
is valid. This method is used internally and not for intended for 
general use.
}
\examples{
check_pattern(input_item$study_id,'ST000001','exact')
}
