% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_def.R
\name{check_puts}
\alias{check_puts}
\alias{check_puts,mw_input_item-method}
\title{Check inputs/outputs match}
\usage{
check_puts(input_item, output_item)

\S4method{check_puts}{mw_input_item}(input_item, output_item)
}
\arguments{
\item{input_item}{An mw_input_item.}

\item{output_item}{An mw_output_item.}
}
\value{
TRUE if the items are compatible or throws an error if not.
}
\description{
Checks that the provided inputs and output are compatible. This method 
is used internally and not for intended for general use.
}
\examples{
check_puts(input_item$study_id,output_item$summary)
}
