% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{context}
\alias{context}
\title{Contexts}
\format{
An object of class \code{list} of length 7.
}
\usage{
context
}
\description{
A predefined list of mw_context objects. The context have been created to
mirror the metabolomics workbench API documentation contexts as closely 
as possible.
}
\examples{
# list available contexts
names(context)

# get the context 'study'
context$study

}
\keyword{datasets}
