% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fcns.R
\name{context_inputs}
\alias{context_inputs}
\title{Valid inputs}
\usage{
context_inputs(context)
}
\arguments{
\item{context}{The name of a valid context (character)}
}
\value{
A list of input item names for a context
}
\description{
Get a list of valid input_items for a context.
}
\examples{
# list of input items for the "study" context
context_inputs("study")
}
