% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fcns.R
\name{context_outputs}
\alias{context_outputs}
\title{Valid outputs}
\usage{
context_outputs(context)
}
\arguments{
\item{context}{The name of a valid context (character)}
}
\value{
A list of output item names for a context
}
\description{
Get a list of valid output_items for a context.
}
\examples{
# list of output items for the "study" context
context_outputs("study")
}
