% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fcns.R
\name{input_example}
\alias{input_example}
\title{Valid input_value for input_item}
\usage{
input_example(input_item)
}
\arguments{
\item{input_item}{The name of a valid input_item (character)}
}
\value{
An example input value matching the pattern required for the chosen
input item.
}
\description{
Displays a valid input_value for an input_item and returns an example that
matches the required input pattern.
}
\examples{
# example input_value for input item "study_id" 
input_example('study_id')
}
