% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomicsWorkbenchR.R
\docType{package}
\name{metabolomicsWorkbenchR}
\alias{metabolomicsWorkbenchR}
\title{metabolomicsWorkbenchR}
\description{
This package provides an interface to the Metabolomics Workbench API. It can 
be used to query the Metabolomics Workbench for study, compound, protein, 
gene information and more. All endpoints of the API are available via a 
simple \code{do_query} method. A number of convenience functions are included
to import study data as \code{SummarizedExperiment} objects to facilitate use
within R and the Bioconductor community.
}
\seealso{
\itemize{
 \item \code{\link{do_query}} for a simple example.
 \item \code{browseVignettes("metabolomicsWorkbenchR")} for more info.
 \item Visit \url{https://www.metabolomicsworkbench.org/tools/mw_rest.php} 
 for details of the API.
}
}
