% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def.R
\name{$,mw_base-method}
\alias{$,mw_base-method}
\title{Get slot value from mw_base objects}
\usage{
\S4method{$}{mw_base}(x, name)
}
\arguments{
\item{x}{An object derived from mw_base.}

\item{name}{The name of the slot to access.}
}
\value{
The assigned to the slot.
}
\description{
Gets the value of a slot from mw_base objects, provided they are not listed as
'private'.
}
\examples{
# an object derived from mw_base object
C = context$study
# access the name slot
C$name
}
