% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{output_item}
\alias{output_item}
\title{Output items}
\format{
An object of class \code{list} of length 28.
}
\usage{
output_item
}
\description{
A predefined list of mw_output_item objects. The items have been created to
mirror the Metabolomics Workbench API documentation output items as closely 
as possible.
}
\examples{
# list available output_items
names(output_item)

# get the output item 'summary'
output_item$summary

}
\keyword{datasets}
