% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_analysis.R
\name{study_analysis}
\alias{study_analysis}
\title{Study analysis}
\usage{
study_analysis(keyword, value)
}
\arguments{
\item{keyword}{the keyword to search by. One of [study_id,study_title,last_name,institute]}

\item{value}{the keyword value. Can be a substring to search by.}
}
\value{
a list
}
\description{
Returns a number of key instrumentation parameters.
}
