% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_data.R
\name{study_data}
\alias{study_data}
\title{Study data}
\usage{
study_data(id)
}
\arguments{
\item{id}{the study or analysis id of a study e.g. "ST000001".}
}
\value{
a data frame
}
\description{
Get the measured values for each metabolite in a study
}
