% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_disease.R
\name{study_disease}
\alias{study_disease}
\title{Study disease}
\usage{
study_disease(study_id)
}
\arguments{
\item{study_id}{the id of a study e.g. "ST000001". Can be a substring to search by.}
}
\value{
a data frame
}
\description{
Get the disease of a study
}
