% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_factors.R
\name{study_factors}
\alias{study_factors}
\title{Study factors}
\usage{
study_factors(study_id)
}
\arguments{
\item{study_id}{the id of a study e.g. "ST000001". Must be a single study id.}
}
\value{
factors for a study
}
\description{
Get factors for a study
}
