% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_metabolites.R
\name{study_metabolites}
\alias{study_metabolites}
\title{Study metabolites}
\usage{
study_metabolites(id)
}
\arguments{
\item{id}{the study or analysis id e.g. "ST000001"}
}
\value{
a list
}
\description{
Details for each named metabolite in a particular study or
analysis
}
