% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_number_metab.R
\name{study_number_metabolites}
\alias{study_number_metabolites}
\title{Study number of metabolites}
\usage{
study_number_metabolites(keyword, value)
}
\arguments{
\item{keyword}{the keyword to search by. One of [study_id,study_title,last_name,institute]}

\item{value}{the keyword value. Can be a substring to search by.}
}
\value{
a data frame
}
\description{
Get the number of metabolites
}
