% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_source.R
\name{study_source}
\alias{study_source}
\title{Study source}
\usage{
study_source(study_id)
}
\arguments{
\item{study_id}{the id of a study e.g. "ST000001". Can be a substring to search by.}
}
\value{
a data frame
}
\description{
Get the source (blood, urine, liver etc) of a study
}
