% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_species.R
\name{study_species}
\alias{study_species}
\title{Study species}
\usage{
study_species(study_id)
}
\arguments{
\item{study_id}{the id of a study e.g. "ST000001". Can be a substring to search by.}
}
\value{
a data frame
}
\description{
Get the species used in a study
}
