% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_summary.R
\name{study_summary}
\alias{study_summary}
\title{Study summary}
\usage{
study_summary(study_id)
}
\arguments{
\item{study_id}{the id of a study e.g. "ST000001". Can be a substring e.g.
"ST" or "ST0004" will return summaries for all studies matching the substring.}
}
\value{
a list of studies and summary information
}
\description{
Get a study summaries
}
