% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomicsWorkbenchR.R
\docType{package}
\name{metabolomicsWorkbenchR}
\alias{metabolomicsWorkbenchR}
\title{metabolomicsWorkbenchR}
\description{
This package provides an interface to the Metabolomics Workbench API. It can 
be used to query the database for study, compound, protein, gene information
and more. All endpoints of the API are availble via simple methods. A number 
of convenience functions are included e.g. to import study data as 
SummarizedExperiment objects to facilite use within R and the Bioconductor 
community.
}
